requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==4.0.1
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
bs4==0.0.2
celery==5.5.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.11
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.0
Flask-Admin==1.6.1
flower==2.0.1
fsspec==2025.3.2
greenlet==3.2.1
h11==0.16.0
huggingface-hub==0.30.2
humanize==4.12.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kombu==5.5.3
langdetect==1.0.9
lxml==5.3.1
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
pyasn1==0.4.8
pydantic==2.11.3
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rsa==4.9.1
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.38
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.6.0
tldextract==5.3.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.1.2
